modificato esercizio

+++ src/main/java/co/develhope/eserciziocrudoperations/controllers/CarController.java
@@ -4,9 +4,12 @@ import co.develhope.eserciziocrudoperations.entities.Car;
 import co.develhope.eserciziocrudoperations.repositories.CarRepo;
 import co.develhope.eserciziocrudoperations.services.CarService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/cars")
@@ -28,19 +31,23 @@ public class CarController {
     }
 
     @GetMapping("/{id}")
-    public Car getCar(@PathVariable long id) {
+    public Optional<Car> getCar(@PathVariable long id) {
         return carService.get(id);
     }
 
     @PutMapping("/{id}")
-    public Car updateCar(@PathVariable long id, @RequestBody Car car) {
-        car.setId(id);
-        return carService.create(car);
+    public Optional<Car> updateCar(@PathVariable long id, @RequestParam String string) {
+        return carService.update(id, string);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteCar(@PathVariable long id) {
-        carService.delete(id);
+    public ResponseEntity<Void> deleteCar(@PathVariable long id) {
+        if (!carRepo.existsById(id)) {
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
+        } else {
+            carService.delete(id);
+            return ResponseEntity.noContent().build();
+        }
     }
 
     @DeleteMapping
